backend/requirements.txt
Django==5.0.2
django-cors-headers==4.3.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
uuid==1.30